Refactor compliance tests into separate focused test files
